pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orbis-modellfabrik"
version = "3.2.0"
description = "ORBIS Modellfabrik Dashboard und Tools"
authors = [{name = "ORBIS Team"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "streamlit>=1.28.0",
    "paho-mqtt>=1.6.1",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.2.0",
]

[project.scripts]
omf2 = "omf2.omf:main"

[tool.setuptools.packages.find]
include = ["omf2*", "session_manager*"]

[tool.setuptools.package-data]
omf2 = [
    "config/*.yml",
    "config/*.json",
    "registry/**/*.yml",
    "registry/**/*.json",
    "assets/**/*.svg",
    "assets/**/*.html",
]

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 120
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (HTML/CSS in f-strings)
]
per-file-ignores = { "**/logger.py" = ["E402"], "**/omf.py" = ["E402"], "integrations/**" = ["F405", "E402"], "session_manager/**" = ["E402"], "tests/**" = ["E402", "F841"] }

[tool.ruff.lint.isort]
known-first-party = ["omf2", "session_manager"]

[tool.pytest.ini_options]
testpaths = ["omf2/tests", "tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
